requirements.txt
flask
flask-cors
python-docx
PyPDF2
python-pptx
rapidfuzz
beautifulsoup4
